fix(S11P12D203-307): Area service 수정

Closes S11P12D203-307

See merge request s11-webmobile1-sub2/S11P12D203!177

+++ backend/mallang/src/main/java/com/chill/mallang/domain/area/dto/ResponseTotalTeamInfo.java
@@ -16,7 +16,7 @@ public class ResponseTotalTeamInfo {
     @NotNull
     private String userName;
     @NotNull
-    private Integer userScore;
+    private Float userScore;
     @NotNull
     private Integer userPlayTime;
 }
